Trim vadmanager code for readableness
